requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
chromadb>=0.5.5
pypdf>=4.3.1
gradio>=4.42.0
sentence-transformers>=3.0.1
python-dotenv>=1.0.1
groq>=0.9.0
pdf2image>=1.17.0
pdfplumber>=0.11.0
Pillow>=10.0.0
ragas>=0.4.0
langchain-groq>=0.1.0
datasets>=2.0.0
